Complete Module 7: Management of workshops and side events

+++ database/migrations/2025_12_03_190232_add_session_id_to_submissions_table.php
@@ -26,8 +26,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('submissions', function (Blueprint $table) {
-            $table->dropConstrainedForeignId('session_id');
+            if (Schema::hasColumn('submissions', 'session_id')) {
             $table->dropColumn('session_id');
+            }
         });
     }
 };
